requirements.txt
tensorflow
Pillow
Flask